fix: auto-kill stale process on bridge port EADDRINUSE (#512)
Move killProcessOnPort logic into BridgeServer.start() so all consumers
(MCP, bench.js, pw-repl-extension) get automatic port cleanup on EADDRINUSE.

Closes #512
